Improve CLI handling by using a library
