+++ services/ssh-portal/main.go
@@ -1,21 +1,12 @@
 package main
 
 import (
-	"context"
-	"encoding/base64"
-	"flag"
 	"fmt"
-	"io"
-	"os"
 
+	"github.com/alecthomas/kong"
 	"github.com/gliderlabs/ssh"
-	"github.com/google/uuid"
 	"github.com/smlx/lagoon/services/ssh-portal/internal/exec"
 	"github.com/smlx/lagoon/services/ssh-portal/internal/keycloak"
-	"github.com/smlx/lagoon/services/ssh-portal/internal/lagoon"
-	lclient "github.com/smlx/lagoon/services/ssh-portal/internal/lagoon/client"
-	"github.com/smlx/lagoon/services/ssh-portal/internal/lagoon/jwt"
-	"github.com/smlx/lagoon/services/ssh-portal/internal/schema"
 	"go.uber.org/zap"
 )
 
@@ -24,27 +15,24 @@ var (
 	buildTime string
 )
 
-type envConfig struct {
-	jwtSecret                string
-	keycloakAuthServerSecret string
-	keycloakBaseURL          string
-	lagoonAPI                string
+// CLI represents the Command Line Interface to ssh-portal
+type CLI struct {
+	Debug                    bool   `kong:"help='enable debug logging'"`
+	Port                     uint   `kong:"default='2222',help='port to listen on'"`
+	JWTSecret                string `kong:"required,env='JWTSECRET',help='Lagoon API JWT secret'"`
+	KeycloakAuthServerSecret string `kong:"required,env='KEYCLOAK_AUTH_SERVER_CLIENT_SECRET',help='auth-server client secret'"`
+	KeycloakAPI              string `kong:"required,env='KEYCLOAK_BASEURL',help='base URL of keycloak API'"`
+	LagoonAPI                string `kong:"required,env='GRAPHQL_ENDPOINT',help='base URL of Lagoon API'"`
 }
 
-// context key types
-type key int
-
-var userKey key
-
 func main() {
-	// parse flags
-	debug := flag.Bool("debug", false, "enable debug logging")
-	port := flag.Int("port", 2222, "listen port")
-	flag.Parse()
+	// parse CLI config
+	cli := CLI{}
+	_ = kong.Parse(&cli)
 	// init logger
 	var log *zap.Logger
 	var err error
-	if *debug {
+	if cli.Debug {
 		log, err = zap.NewDevelopment()
 	} else {
 		log, err = zap.NewProduction()
@@ -53,153 +41,28 @@ func main() {
 		panic(err)
 	}
 	defer log.Sync()
-
 	log.Info("startup",
 		zap.String("version", version), zap.String("buildTime", buildTime))
-
-	// get environmental configuration
-	config, err := getEnvConfig()
-	if err != nil {
-		log.Fatal("couldn't get environmental configuration", zap.Error(err))
-	}
-
-	k, err := keycloak.New(config.keycloakBaseURL,
-		config.keycloakAuthServerSecret, log)
+	// init keycloak client
+	k, err := keycloak.New(cli.KeycloakAPI, cli.KeycloakAuthServerSecret, log)
 	if err != nil {
 		log.Fatal("couldn't get keycloak client", zap.Error(err))
 	}
-
+	// init k8s exec client
 	e, err := exec.New()
 	if err != nil {
 		log.Fatal("couldn't get exec client", zap.Error(err))
 	}
-
 	// configure ssh connection handling
-	ssh.Handle(sessionHandler(k, e, config.lagoonAPI, config.jwtSecret, log, *debug))
-
+	ssh.Handle(sessionHandler(k, e, cli.LagoonAPI, cli.JWTSecret, log, cli.Debug))
+	// start SSH server
 	log.Fatal("server error", zap.Error(
 		ssh.ListenAndServe(
-			fmt.Sprintf(":%d", *port),
+			fmt.Sprintf(":%d", cli.Port),
 			nil,
 			ssh.PublicKeyAuth(pubKeyAuth(
 				log,
-				config.jwtSecret,
-				config.lagoonAPI,
-				*debug)))))
-}
-
-// pubKeyAuth performs a check against the lagoon API for the public key
-func pubKeyAuth(log *zap.Logger, jwtSecret, lagoonAPI string, debug bool) ssh.PublicKeyHandler {
-	return func(ctx ssh.Context, key ssh.PublicKey) bool {
-		// generate a JWT token
-		token, err := jwt.OneMinuteAdminToken(jwtSecret)
-		if err != nil {
-			log.Error("couldn't get JWT token", zap.Error(err))
-			return false
-		}
-		// get the lagoon client with the admin token
-		l := lclient.New(lagoonAPI, token, "ssh-portal "+version, debug)
-		// get the user ID from lagoon
-		keyLogField := zap.String("publicKey", fmt.Sprintf("%s %s",
-			key.Type(),
-			base64.StdEncoding.EncodeToString(key.Marshal())))
-		user, err := lagoon.UserBySSHKey(context.TODO(), l, key)
-		if err != nil {
-			log.Debug("unknown SSH key", keyLogField, zap.Error(err))
-			return false
-		}
-		log.Info("accepted public key", keyLogField, zap.String("userID", user.ID.String()))
-		ctx.SetValue(userKey, user)
-		return true
-	}
-}
-
-// sessionHandler contains the main ssh session logic
-func sessionHandler(k *keycloak.Client, c *exec.Client,
-	lagoonAPI, jwtSecret string, log *zap.Logger, debug bool) ssh.Handler {
-	return func(s ssh.Session) {
-		// generate session ID
-		sid := uuid.New()
-		log.Info("start session", zap.String("sessionID", sid.String()))
-		defer log.Info("end session", zap.String("sessionID", sid.String()))
-		// extract the user object that was added to the context during
-		// authentication
-		user, ok := s.Context().Value(userKey).(*schema.User)
-		if !ok {
-			log.Error("unknown context value for user",
-				zap.String("sessionID", sid.String()))
-			io.WriteString(s, "internal error\n")
-			return
-		}
-		log.Info("identified user",
-			zap.String("userID", user.ID.String()),
-			zap.String("userEmail", user.Email),
-			zap.String("sessionID", sid.String()))
-		// get a user token from keycloak
-		ctoken, err := k.UserToken(&user.ID)
-		if err != nil {
-			log.Warn("couldn't get user token", zap.Error(err),
-				zap.String("sessionID", sid.String()))
-			io.WriteString(s, "internal error\n")
-			return
-		}
-		// check the user and command. if it is "lagoon" and "token" respectively,
-		// return the user token and exit. This is used to get a graphql token via
-		// SSH.
-		cmd := s.Command()
-		if s.User() == "lagoon" && len(cmd) == 1 && cmd[0] == "token" {
-			log.Info("issuing user token", zap.String("sessionID", sid.String()))
-			io.WriteString(s, ctoken)
-			return
-		}
-		// get the lagoon client using the user token
-		cl := lclient.New(lagoonAPI, ctoken, "ssh-portal "+version, true)
-		// Now, authenticated as the user, check for SSH permissions on the
-		// namespace. Here, s.User() is the ssh username - for Lagoon this is the
-		// namespace name
-		canSSH, err := lagoon.UserCanSSHToEnvironment(context.TODO(), cl, s.User())
-		if err != nil {
-			log.Warn("couldn't get user SSH permissions", zap.Error(err),
-				zap.String("sessionID", sid.String()))
-			io.WriteString(s, "internal error\n")
-			return
-		}
-		if !canSSH {
-			log.Info("permission denied", zap.Error(err),
-				zap.String("sessionID", sid.String()))
-			io.WriteString(s, "permission denied\n")
-			return
-		}
-		// check if a pty is required
-		_, _, pty := s.Pty()
-		// start the command
-		err = c.Exec("cli", s.User(), s.Command(), s, s.Stderr(), pty)
-		if err != nil {
-			log.Warn("couldn't execute command", zap.Error(err),
-				zap.String("sessionID", sid.String()))
-			io.WriteString(s, "couldn't execute command\n")
-		}
-	}
-}
-
-func getEnvConfig() (*envConfig, error) {
-	config := envConfig{}
-	config.lagoonAPI = os.Getenv("GRAPHQL_ENDPOINT")
-	if len(config.lagoonAPI) == 0 {
-		return &config, fmt.Errorf("GRAPHQL_ENDPOINT not set")
-	}
-	config.keycloakBaseURL = os.Getenv("KEYCLOAK_BASEURL")
-	if len(config.keycloakBaseURL) == 0 {
-		return &config, fmt.Errorf("KEYCLOAK_BASEURL not set")
-	}
-	config.keycloakAuthServerSecret =
-		os.Getenv("KEYCLOAK_AUTH_SERVER_CLIENT_SECRET")
-	if len(config.keycloakAuthServerSecret) == 0 {
-		return &config, fmt.Errorf("KEYCLOAK_AUTH_SERVER_CLIENT_SECRET not set")
-	}
-	config.jwtSecret = os.Getenv("JWTSECRET")
-	if len(config.jwtSecret) == 0 {
-		return &config, fmt.Errorf("JWTSECRET not set")
-	}
-	return &config, nil
+				cli.JWTSecret,
+				cli.LagoonAPI,
+				cli.Debug)))))
 }
